refactor: rename v2 utils and types files
