+++ src/hooks/use-has-service-consent.ts
@@ -1,8 +1,8 @@
 import { useContext } from 'react'
 
 import { UsercentricsContext } from '../context.js'
-import type { ServiceId } from '../types.js'
-import { hasServiceConsent } from '../utils.js'
+import type { ServiceId } from '../types.v2.js'
+import { hasServiceConsent } from '../utils.v2.js'
 import { useServiceDebug } from './use-service-debug.js'
 import { useServiceInfo } from './use-service-info.js'
