+++ src/hooks/use-service-info.ts
@@ -1,8 +1,8 @@
 import { useContext, useEffect, useState } from 'react'
 
 import { UsercentricsContext } from '../context.js'
-import type { ServiceFullInfo, ServiceId, ServiceInfo } from '../types.js'
-import { getServicesBaseInfo, getServicesFullInfo } from '../utils.js'
+import type { ServiceFullInfo, ServiceId, ServiceInfo } from '../types.v2.js'
+import { getServicesBaseInfo, getServicesFullInfo } from '../utils.v2.js'
 import { useServiceDebug } from './use-service-debug.js'
 
 /**
