+++ src/index.ts
@@ -8,5 +8,5 @@ export { useIsFailed } from './hooks/use-is-failed.js'
 export { useIsInitialized } from './hooks/use-is-initialized.js'
 export { useIsOpen } from './hooks/use-is-open.js'
 export { useServiceFullInfo, useServiceInfo } from './hooks/use-service-info.js'
-export * from './types.js'
-export * from './utils.js'
+export * from './types.v2.js'
+export * from './utils.v2.js'
